requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
httptools==0.6.4
idna==3.13
iniconfig==2.1.0
packaging==26.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-core==2.27.2
pypdf==5.9.0
pytest==8.3.5
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.44.0
tomli==2.4.1
typing-extensions==4.13.2
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
